pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "simplifier"
version = "0.1.0"
description = 'Simple Simplify German Language'
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"


authors = [
  { name = "machinelearningZH", email = "datashop@statistik.zh.ch" },
]

dependencies = [
    "ipywidgets>=8.1.5",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.0.1",
    "openai>=1.58.1",
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "jsonpath-ng>=1.7.0",
    "pydantic>=2.10.5",
    "itsdangerous>=2.2.0",
]

[project.urls]
Documentation = "https://github.com/machinelearningZH/simple-simplifier#readme"
Issues = "https://github.com/machinelearningZH/simple-simplifier/issues"
Source = "https://github.com/machinelearningZH/simple-simplifier"

[tool.hatch]
version = "0.1.0"

[dependency-groups]
dev = [
    "ruff>=0.8.5",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Ignore line length errors (especially for long strings/comments)